requirements.txt
streamlit==1.32.0
rdkit-pypi==2022.9.5
py3Dmol==2.0.4
stmol==0.1.2
